+++ packages/terra-framework-docs/src/terra-dev-site/test/collapsible-menu-view/CollapsibleMenuViewSingleItem.test.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import CollapsibleMenuView from '../../../CollapsibleMenuView';
+import CollapsibleMenuView from 'terra-collapsible-menu-view';
 
 const CollapsibleMenuViewSingleItem = () => (
   <CollapsibleMenuView>