+++ packages/terra-framework-docs/src/terra-dev-site/test/collapsible-menu-view/FirstHiddenItemIsIcon.test.jsx
@@ -1,13 +1,9 @@
 import React from 'react';
 import IconSend from 'terra-icon/lib/icon/IconSend';
-import CollapsibleMenuView from '../../../CollapsibleMenuView';
+import CollapsibleMenuView from 'terra-collapsible-menu-view';
 
 const CollapsibleMenuViewDemo = () => (
-  <CollapsibleMenuView alwaysCollapsedMenuItems={[
-    <CollapsibleMenuView.Item text="Collapsed Button 1" key="collapsedButton1" />,
-    <CollapsibleMenuView.Item text="Collapsed Button 2" key="collapsedButton2" />,
-    <CollapsibleMenuView.Item text="Collapsed Button 3" key="collapsedButton3" />]}
-  >
+  <CollapsibleMenuView>
     <CollapsibleMenuView.Item text="Button A" className="ButtonA" key="buttonA" />
     <CollapsibleMenuView.Item text="Button B" className="ButtonB" key="buttonB" />
     <CollapsibleMenuView.Divider key="Divider4" />